requirements.txt
moviepy
azure-cognitiveservices-speech
